requirements.txt
Django==1.11.10
django-taggit==0.22.2
Markdown==2.6.11
pytz==2018.4
